Add TAOS-CI (Opensource-friendly CI system) to tools section
TAOS-CI (Apache 2.0) is a GitHub webhook API based lightweight & modular
continuous integration system for "Review less, merge faster".
* https://github.com/nnsuite/TAOS-CI

It is an automated project coordinator to prevent a regression, find bugs,
and reduce a nonproductive review process because of incorrect PRs in
github.com. Actually, Submitting the incorrect PRs is a PITA in case of
continuous integration. Basically, PRs causing regressions will not be
automatically merged. As a result of that, it will reduce the burdens of
reviewers.

* Minimize a nonproductive review process
* Provide a test automation (both build and run)
* Prevent a performance regression
* Find bugs at a proper time before merging buggy codes
* Generate a doxygen-based developer manual
* Support modulable facilities with plug-in interface
* Integrate the existing opensource tools easily
* Verify an integrity of a package by supporting a platform build
* Support multiple operating systems

**Changes proposed in this PR:**
1. Added TAOS-CI to tools seciton

Signed-off-by: Geunsik Lim <leemgs@gmail.com>